Add shot segmentation for keyframe extraction
For each I-frame's blue, green, and red channels, a histogram is computed. Then each histogram is compared between
frames--if the correlation is lower than a threshold, then the frames belong to different shots and are returned
as keyframes.

This should be made more efficient in the future by saving the histogram of keyframe[i+1] instead of calculating it twice.